Use controller and LinkGenerator for anime posters
Also:
 * Remove static files provider for images
 * Add IContentTypeProvider DI
