+++ Shizou.Blazor/Features/Anime/Anime.razor
@@ -1,4 +1,5 @@
 ﻿@page "/Collection/{AnimeId:int}"
+@using Shizou.Server.Controllers
 
 <article class="container-fluid">
     @if (_anime is not null)
@@ -6,9 +7,10 @@
         <PageTitle>@_anime.TitleTranscription</PageTitle>
         <div class="row justify-content-center">
             <div class="col-xl-3 col-lg-4 col-md-6 col-12 d-flex justify-content-center justify-content-md-end anime-poster">
+                @* ReSharper disable once RedundantAnonymousTypePropertyName *@
                 @if (_anime.ImageFilename is not null)
                 {
-                    <img src="@($"{WebPaths.AnimePostersDir}/{_anime.ImageFilename}")">
+                    <img src="@(LinkGenerator.GetPathByAction(nameof(Images.GetAnimePoster), nameof(Images), new { AnimeId = AnimeId }))">
                 }
             </div>
             <div class="col-xl-3 col-lg-4 col-md-6 col-12 d-flex justify-content-start align-items-start">
