+++ Shizou.Blazor/Features/Components/AnimeCard.razor
@@ -1,7 +1,8 @@
-﻿<div class="card anime-card m-2" role="button" @onclick="@(() => GoToAnime(AniDbAnime.Id))">
+﻿@using Shizou.Server.Controllers
+<div class="card anime-card m-2" role="button" @onclick="@(() => GoToAnime(AniDbAnime.Id))">
     @if (AniDbAnime.ImageFilename is not null)
     {
-        <img class="card-img object-fit-contain h-100" src="@($"{WebPaths.AnimePostersDir}/{AniDbAnime.ImageFilename}")">
+        <img class="card-img object-fit-contain h-100" src="@(LinkGenerator.GetPathByAction(nameof(Images.GetAnimePoster), nameof(Images), new { AnimeId = AniDbAnime.Id }))">
     }
     <div class="card-img-overlay d-flex p-0 align-items-end">
         <div class="p-2 text-center text-white bg-black bg-opacity-25 w-100 overflow-hidden" style="text-shadow: 2px -1px 2px #000">
