--- Shizou.Server/Controllers/ImagesController.cs
@@ -1,27 +1,0 @@
-﻿using Microsoft.AspNetCore.Http;
-using Microsoft.AspNetCore.Http.HttpResults;
-using Microsoft.AspNetCore.Mvc;
-using Shizou.Server.Services;
-using Swashbuckle.AspNetCore.Annotations;
-
-namespace Shizou.Server.Controllers;
-
-[ApiController]
-[Route("api/[controller]")]
-public class ImagesController : ControllerBase
-{
-    private readonly ImageService _imageService;
-
-    public ImagesController(ImageService imageService)
-    {
-        _imageService = imageService;
-    }
-
-    [HttpPut("[action]")]
-    [SwaggerResponse(StatusCodes.Status200OK)]
-    public Ok GetMissingAnimePosters()
-    {
-        _imageService.GetMissingAnimePosters();
-        return TypedResults.Ok();
-    }
-}
